Update "Ruby gems" article with aliases
We've recently started using ruby alias versions in R&D to mitigate
issues with versions on our CI where only the most recent patch
version is available.

This modification to the "The better way with Ruby Gems" article
introduces the concept of Ruby version aliases and how to configure
them, as well as touching on the advantages / disadvantages of taking
this route.